work in progress
Almost there. it now prevents the 3 categories of interactions. but it also freaks out when you interact with something not in the list. Shabbat is starting. I'll finish knocking this out on Yom Aleph.

When releases this will feel like an upgrade as it now has a client side config. and because it's 1.7.10 I can port it forward to 1.8, 1.9.4, 1.10.2, etc.